fix Intel regions backup
We have detected that the ME and IGB regions are not backed up
on MTL laptops (should affect other hw as well) due to the
mismatch in this detection logic.

Another issue is why we have two functions doing the same thing,
but this can be resolved by a followup commits.

Signed-off-by: Maciej Pijanowski <maciej.pijanowski@3mdeb.com>