include: check for firmware bianaries instead of credentials
Previous function check_des_creds was designed to verify both platform
compatibility and firmware binaries availability. If the platform was
compatible but the credentials did not have access to the firmware
binaries - the function will state that the credentials are corrupted
and save standart credentials instead.

It is not a case anymore, because after introducing package management
system the credentials may not have access to the firmware but to some
packages, so it is not correct to make any assumptions on the credentials
while checking for firmware.

Signed-off-by: Daniil Klimuk <daniil.klimuk@3mdeb.com>